python-app/anomaly-detection/pyproject.toml
[tool.poetry]
name = "anomaly-detection"
version = "0.1.0"
description = ""
authors = ["Emanuele Fumagalli <emanuelef@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
compress = "^0.0.3"
statsforecast = "^1.7.3"
httpx = "^0.27.0"
pandas = "^2.2.1"
prophet = "^1.1.5"
cachetools = "^5.3.3"
opentelemetry-instrumentation-fastapi = "^0.45b0"
python-dotenv = "^1.0.1"
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-semantic-conventions = "^0.45b0"
opentelemetry-exporter-otlp-proto-grpc = "^1.24.0"
opentelemetry-instrumentation-httpx = "^0.45b0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
